Issue #652: New CLI option to print Javadoc comment parse tree
