+++ config/suppressions.xml
@@ -129,7 +129,7 @@
     <!-- we need that set of converters -->
     <suppress checks="ClassDataAbstractionCoupling" files="AutomaticBean\.java"/>
     <!-- they are aggregators of logic, usage a several of classes are ok -->
-    <suppress checks="ClassDataAbstractionCoupling" files="(Checker|TreeWalker|Main|CheckstyleAntTask|AbstractJavadocCheck)\.java"/>
+    <suppress checks="ClassDataAbstractionCoupling" files="(Checker|TreeWalker|Main|CheckstyleAntTask|JavadocDetailNodeParser)\.java"/>
     <suppress checks="ClassDataAbstractionCoupling" files="(CheckerTest|TreeWalkerTest|BaseCheckTestSupport|XDocsPagesTest|CheckstyleAntTaskTest)\.java"/>
     <!-- a lot of GUI elements is OK -->
     <suppress checks="ClassDataAbstractionCoupling" files="(JTreeTable|MainFrame)\.java"/>
@@ -143,7 +143,7 @@
 
     <!-- Suppressions from PMD configuration-->
     <!-- validateCli is not reasonable to split as encapsulation of logic will be damaged -->
-    <suppress checks="CyclomaticComplexity" files="Main\.java"  lines="200"/>
+    <suppress checks="CyclomaticComplexity" files="Main\.java"  lines="211"/>
     <!-- JavadocMethodCheck, JavadocStyleCheck, JavadocUtils.getJavadocTags() - deprecated -->
     <suppress checks="CyclomaticComplexity" files="JavadocMethodCheck\.java"/>
     <suppress checks="CyclomaticComplexity" files="JavadocStyleCheck\.java"/>
@@ -164,6 +164,7 @@
               lines="105, 143, 174"/>
 
     <!-- Not reasonable to split as encapsulation of logic will be damaged. -->
-    <suppress checks="ExecutableStatementCount" files="Main\.java" lines="102"/>
-    <suppress checks="CyclomaticComplexity" files="Main\.java"  lines="102"/>
+    <suppress checks="ExecutableStatementCount" files="Main\.java" lines="108"/>
+    <suppress checks="CyclomaticComplexity" files="Main\.java"  lines="108"/>
+    <suppress checks="JavaNCSS" files="Main\.java"  lines="108"/>
 </suppressions>
