+++ src/test/resources/com/puppycrawl/tools/checkstyle/astprinter/InputJavadocComment.javadoc
@@ -1,0 +1,4 @@
+<p>
+My {@code Class}
+</p>
+@see never
