+++ src/test/resources/com/puppycrawl/tools/checkstyle/astprinter/expectedInputJavadocComment.txt
@@ -1,0 +1,36 @@
+JAVADOC -> <p>\r\nMy {@code Class}\r\n</p>\r\n@see never<EOF> [0:0]
+|--HTML_ELEMENT -> <p>\r\nMy {@code Class}\r\n</p> [0:0]
+|   `--PARAGRAPH -> <p>\r\nMy {@code Class}\r\n</p> [0:0]
+|       |--P_TAG_OPEN -> <p> [0:0]
+|       |   |--OPEN -> < [0:0]
+|       |   |--P_HTML_TAG_NAME -> p [0:1]
+|       |   `--CLOSE -> > [0:2]
+|       |--NEWLINE -> \r\n [0:3]
+|       |--TEXT -> My  [1:0]
+|       |   |--CHAR -> M [1:0]
+|       |   |--CHAR -> y [1:1]
+|       |   `--WS ->   [1:2]
+|       |--JAVADOC_INLINE_TAG -> {@code Class} [1:3]
+|       |   |--JAVADOC_INLINE_TAG_START -> { [1:3]
+|       |   |--CODE_LITERAL -> @code [1:4]
+|       |   |--WS ->   [1:9]
+|       |   |--TEXT -> Class [1:10]
+|       |   |   |--CHAR -> C [1:10]
+|       |   |   |--CHAR -> l [1:11]
+|       |   |   |--CHAR -> a [1:12]
+|       |   |   |--CHAR -> s [1:13]
+|       |   |   `--CHAR -> s [1:14]
+|       |   `--JAVADOC_INLINE_TAG_END -> } [1:16]
+|       |--NEWLINE -> \r\n [1:17]
+|       `--P_TAG_CLOSE -> </p> [2:0]
+|           |--OPEN -> < [2:0]
+|           |--SLASH -> / [2:1]
+|           |--P_HTML_TAG_NAME -> p [2:2]
+|           `--CLOSE -> > [2:3]
+|--NEWLINE -> \r\n [2:4]
+|--JAVADOC_TAG -> @see never [3:0]
+|   |--SEE_LITERAL -> @see [3:0]
+|   |--WS ->   [3:4]
+|   `--REFERENCE -> never [3:5]
+|       `--PACKAGE -> never [3:5]
+`--EOF -> <EOF> [3:10]
